fixing some things
* items are added to the 'previous' tables when settings are saved
* jobs now save if a freq job is for numerical frequencies
* presets can also save this (check the job type box)
* cannot kill a completed job
* added opening the queue tool under view on the input builder

+++ src/tools/job_manager_tool.py
@@ -170,8 +170,9 @@ class JobQueue(ToolInstance):
         ndxs = list(set([item.row() for item in self.tree.selectedIndexes()]))
         for ndx in ndxs:
             job = jobs[ndx]
-            job.terminate()
-            job.wait()
+            if not job.isFinished():
+                job.terminate()
+                job.wait()
             
         self.session.seqcrow_job_manager.triggers.activate_trigger(JOB_QUEUED, "kill")
